Move gene panel rule to case rules
