+++ cg/services/orders/validation/order_types/mip_dna/validation_rules.py
@@ -8,6 +8,7 @@ from cg.services.orders.validation.rules.case.rules import (
     validate_each_new_case_has_an_affected_sample,
     validate_existing_cases_belong_to_collaboration,
     validate_existing_cases_have_an_affected_sample,
+    validate_gene_panels_exist,
     validate_gene_panels_unique,
     validate_samples_in_case_have_same_prep_category,
 )
@@ -24,7 +25,6 @@ from cg.services.orders.validation.rules.case_sample.rules import (
     validate_existing_samples_compatible_with_order_type,
     validate_fathers_are_male,
     validate_fathers_in_same_case_as_children,
-    validate_gene_panels_exist,
     validate_mothers_are_female,
     validate_mothers_in_same_case_as_children,
     validate_not_all_samples_unknown_in_case,
